backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment Variables
python-dotenv>=1.0,<2.0

# File Handling
Pillow>=10.0,<11.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Development
django-debug-toolbar>=4.2,<5.0

# Email (optional, for contact form notifications)
django-anymail>=10.0,<11.0

